pyproject.toml
[tool.poetry]
name = "datasets-preview-backend"
version = "0.4.6"
description = "API to extract rows of 🤗 datasets"
authors = ["Sylvain Lesage <severo@rednegra.net>"]

[tool.poetry.dependencies]
python = "^3.8"
starlette = "^0.16.0"
uvicorn = "^0.14.0"
Pillow = "^8.3.1"
trec-car-tools = {path = "vendors/trec-car-tools/python3"}
apache-beam = "^2.31.0"
conllu = "^4.4"
kss = "^2.5.1"
lm-dataformat = "^0.0.19"
lxml = "^4.6.3"
nlp = "^0.4.0"
openpyxl = "^3.0.7"
py7zr = "^0.16.1"
tensorflow = "^2.5.0"
wget = "^3.2"
kenlm = {url = "https://github.com/kpu/kenlm/archive/master.zip"}
nltk = "^3.6.2"
aiohttp = "^3.7.4"
typer = "^0.3.2"
transformers = "^4.10.0"
datasets = {git = "https://github.com/huggingface/datasets.git", rev = "a67992906ebe3f81e8047c58ba3392d2ec89fb67"}
python-dotenv = "^0.19.0"
watchdog = {extras = ["watchmedo"], version = "^2.1.3"}
diskcache = "^5.2.1"

[tool.poetry.dev-dependencies]
bandit = "^1.7.0"
black = "21.7b0"
flake8 = "^3.9.2"
isort = "^5.9.3"
pytest = "^6.2.4"
safety = "^1.10.3"
mypy = "^0.910"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
filterwarnings = [
    "ignore::DeprecationWarning"
]

[tool.isort]
profile = "black"

[tool.black]
line-length = 119
experimental-string-processing = true

[tool.mypy]
strict = true

[[tool.mypy.overrides]]
# could be solved in datasets by adding __add__ to the __init__.py file
module = "datasets"
implicit_reexport = true
